TRIPLEX-852 Notification: AI-Ready (forwardRef, AI.md, modern stories)
