+++ src/components/Notification/NotificationGrouped.tsx
@@ -2,22 +2,24 @@ import { NotificationGroupedFooter } from "./components/NotificationGroupedFoote
 import React from "react";
 import styles from "./styles/Notification.module.less";
 
-/**
- * Свойства NotificationGrouped.
- */
+/** Свойства NotificationGrouped. */
 export interface INotificationGroupedProps {
+    /** Нотификация. */
     children: React.ReactNode;
 }
 
 /**
  * Компонент NotificationGrouped.
- * @prop {React.ReactNode} children Нотификация.
  */
-export const NotificationGrouped: React.FC<INotificationGroupedProps> = ({ children }) => (
-    <div className={styles.notificationGroupedWrapper}>
-        {children}
-        <NotificationGroupedFooter />
-    </div>
+export const NotificationGrouped = React.forwardRef<HTMLDivElement, INotificationGroupedProps>(
+    function NotificationGrouped({ children }, ref) {
+        return (
+            <div className={styles.notificationGroupedWrapper} ref={ref}>
+                {children}
+                <NotificationGroupedFooter />
+            </div>
+        );
+    },
 );
 
 NotificationGrouped.displayName = "NotificationGrouped";
