+++ src/components/Notification/components/NotificationBodyContent.tsx
@@ -4,6 +4,7 @@ import { ETextSize, EFontType } from "../../Typography/enums";
 
 /** Свойства компонента NotificationBodyContent. */
 interface INotificationBodyContentProps {
+    /** Основное сообщение нотификации. */
     children?: React.ReactNode;
 }
