+++ src/components/Notification/components/NotificationBodyList.tsx
@@ -5,6 +5,7 @@ import clsx from "clsx";
 
 /** Свойства компонента NotificationBodyList. */
 interface INotificationBodyListProps extends IUnorderedListProps {
+    /** Значения элементов списка. */
     values: string[];
 }
