+++ src/components/Notification/components/NotificationHeader.tsx
@@ -1,12 +1,9 @@
 import React from "react";
 import styles from "../styles/Notification.module.less";
 
-/**
- * Опции хэдера нотификации.
- *
- * @prop {React.ReactNode} children Дочерний элемент хэдера нотификации.
- */
+/** Опции хэдера нотификации. */
 export interface INotificationHeaderProps {
+    /** Дочерний элемент хэдера нотификации. */
     children: React.ReactNode;
 }
