+++ stories/Notification/Notification.stories.tsx
@@ -1,32 +1,28 @@
-import { Controls, Description, Primary, Stories, Subtitle, Title } from "@storybook/addon-docs/blocks";
-import { Notification } from "../../src/components/Notification/Notification";
-import { NotificationGrouped } from "../../src/components/Notification/NotificationGrouped";
+import React from "react";
+import { Meta, StoryObj } from "@storybook/react";
+import { ArgTypes, Controls, Description, Heading, Primary, Stories, Title } from "@storybook/addon-docs/blocks";
+import { Notification } from "@sberbusiness/triplex-next";
 import {
-    DefaulticonStrokePrdIcon20,
-    ErrorStrokeStsIcon20,
-    SuccessStrokeStsIcon20,
-    WarningStrokeStsIcon20,
-    SuccessgradientStsIcon96,
-    StarStrokeSrvIcon32,
-} from "@sberbusiness/icons-next";
-import { StoryObj } from "@storybook/react";
-import { action } from "storybook/actions";
-import { EButtonTheme, Button } from "../../src/components/Button";
-import { EComponentSize } from "../../src/enums/EComponentSize";
-import { AlertContext, EAlertType } from "../../src/components/Alert";
-import React, { useState } from "react";
-import { EFormFieldStatus } from "../../src/components/FormField";
-import { Text, ETextSize, EFontType } from "../../src/components/Typography";
-import { TextField } from "../../src/components/TextField";
-import { Checkbox, CheckboxYGroup } from "../../src/components/Checkbox";
+    Business as BusinessRender,
+    BusinessSource,
+    BusinessStack as BusinessStackRender,
+    BusinessStackSource,
+    FeedbackWithoutStars as FeedbackWithoutStarsRender,
+    FeedbackWithoutStarsSource,
+    FeedbackWithStars as FeedbackWithStarsRender,
+    FeedbackWithStarsSource,
+    INotificationPlaygroundProps,
+    Playground as PlaygroundRender,
+    Status as StatusRender,
+    StatusSource,
+} from "./examples";
 import "./Notification.less";
 
-export default {
+const meta = {
     title: "Components/Notification",
     component: Notification,
     tags: ["autodocs"],
     parameters: {
-        testRunner: { skip: true },
         docs: {
             description: {
                 component:
@@ -35,8 +31,10 @@ export default {
             page: () => (
                 <>
                     <Title />
-                    <Subtitle />
                     <Description />
+                    <Heading>Props</Heading>
+                    <ArgTypes of={Notification} />
+                    <Heading>Playground</Heading>
                     <Primary />
                     <Controls of={Playground} />
                     <Stories />
@@ -44,27 +42,12 @@ export default {
             ),
         },
     },
-};
+} satisfies Meta<typeof Notification>;
 
-interface INotificationPlaygroundProps {
-    withExtraBottomPadding: boolean;
-    isShowCloseOnHover: boolean;
-    showIcon: boolean;
-    iconType: "success" | "warning" | "error" | "default";
-    showHeader: boolean;
-    headerText: string;
-    showContent: boolean;
-    contentText: string;
-    showList: boolean;
-    listItems: string;
-    showFooter: boolean;
-    showClose: boolean;
-    showTime: boolean;
-    time: string;
-}
+export default meta;
 
 export const Playground: StoryObj<INotificationPlaygroundProps> = {
-    name: "Playground",
+    tags: ["!autodocs"],
     args: {
         withExtraBottomPadding: false,
         isShowCloseOnHover: false,
@@ -104,6 +87,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
+                category: "Settings",
             },
         },
         iconType: {
@@ -113,6 +97,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "success | warning | error | default" },
                 defaultValue: { summary: "success" },
+                category: "Settings",
             },
         },
         showHeader: {
@@ -121,6 +106,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
+                category: "Settings",
             },
         },
         headerText: {
@@ -128,6 +114,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             description: "Текст заголовка",
             table: {
                 type: { summary: "string" },
+                category: "Settings",
             },
         },
         showContent: {
@@ -136,6 +123,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
+                category: "Settings",
             },
         },
         contentText: {
@@ -143,6 +131,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             description: "Текст содержимого",
             table: {
                 type: { summary: "string" },
+                category: "Settings",
             },
         },
         showList: {
@@ -151,6 +140,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "false" },
+                category: "Settings",
             },
         },
         listItems: {
@@ -158,6 +148,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             description: "Элементы списка (разделенные точкой с запятой)",
             table: {
                 type: { summary: "string" },
+                category: "Settings",
             },
         },
         showFooter: {
@@ -166,6 +157,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "false" },
+                category: "Settings",
             },
         },
         showClose: {
@@ -174,6 +166,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "true" },
+                category: "Settings",
             },
         },
         showTime: {
@@ -182,6 +175,7 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             table: {
                 type: { summary: "boolean" },
                 defaultValue: { summary: "false" },
+                category: "Settings",
             },
         },
         time: {
@@ -189,568 +183,83 @@ export const Playground: StoryObj<INotificationPlaygroundProps> = {
             description: "Время",
             table: {
                 type: { summary: "string" },
+                category: "Settings",
             },
         },
     },
     parameters: {
-        controls: {
-            include: [
-                "withExtraBottomPadding",
-                "isShowCloseOnHover",
-                "showIcon",
-                "iconType",
-                "showHeader",
-                "headerText",
-                "showContent",
-                "contentText",
-                "showList",
-                "listItems",
-                "showFooter",
-                "showClose",
-                "showTime",
-                "time",
-            ],
-        },
+        testRunner: { skip: true },
         docs: {
-            description: {
-                story: "Интерактивная демонстрация Notification с расширенными controls. Позволяет настраивать все основные свойства компонента, включая отображение иконки, заголовка, содержимого, списка, футера, кнопки закрытия и времени.",
-            },
+            canvas: { sourceState: "none" },
         },
     },
-    render: (args) => {
-        const getIcon = () => {
-            switch (args.iconType) {
-                case "success":
-                    return <SuccessStrokeStsIcon20 paletteIndex={0} />;
-                case "warning":
-                    return <WarningStrokeStsIcon20 paletteIndex={2} />;
-                case "error":
-                    return <ErrorStrokeStsIcon20 paletteIndex={1} />;
-                case "default":
-                    return <DefaulticonStrokePrdIcon20 paletteIndex={5} />;
-                default:
-                    return <SuccessStrokeStsIcon20 paletteIndex={0} />;
-            }
-        };
-
-        const listValues = args.listItems ? args.listItems.split(";").filter((item) => item.trim() !== "") : [];
-
-        const children: React.ReactElement[] = [];
-
-        if (args.showIcon) {
-            children.push(<Notification.Icon key="icon">{getIcon()}</Notification.Icon>);
-        }
-
-        const bodyChildren: React.ReactElement[] = [];
-        if (args.showHeader) {
-            bodyChildren.push(<Notification.Body.Header key="header">{args.headerText}</Notification.Body.Header>);
-        }
-        if (args.showContent) {
-            bodyChildren.push(<Notification.Body.Content key="content">{args.contentText}</Notification.Body.Content>);
-        }
-        if (args.showList && listValues.length > 0) {
-            bodyChildren.push(<Notification.Body.List key="list" values={listValues} />);
-        }
-        if (args.showFooter) {
-            bodyChildren.push(<Notification.Body.Footer key="footer">Footer text</Notification.Body.Footer>);
-        }
-
-        children.push(<Notification.Body key="body">{bodyChildren}</Notification.Body>);
-
-        if (args.showClose) {
-            children.push(<Notification.Close key="close" onClick={action("onClose")} />);
-        }
-
-        if (args.showTime) {
-            children.push(<Notification.Time key="time" time={args.time} />);
-        }
-
-        return (
-            <div style={{ maxWidth: "600px" }}>
-                <Notification
-                    withExtraBottomPadding={args.withExtraBottomPadding}
-                    isShowCloseOnHover={args.isShowCloseOnHover}
-                    onClick={action("onClick")}
-                >
-                    {children}
-                </Notification>
-            </div>
-        );
-    },
+    render: (args) => <PlaygroundRender {...args} />,
 };
 
 export const Status: StoryObj<typeof Notification> = {
-    name: "Status",
+    render: StatusRender,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: "32px" }}>
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Success</h3>
-                    <Notification>
-                        <Notification.Icon>
-                            <SuccessStrokeStsIcon20 paletteIndex={0} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                            <Notification.Body.List values={Array.from({ length: 3 }, () => "List item text;")} />
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Warning</h3>
-                    <Notification>
-                        <Notification.Icon>
-                            <WarningStrokeStsIcon20 paletteIndex={2} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                            <Notification.Body.List values={Array.from({ length: 3 }, () => "List item text;")} />
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Error</h3>
-                    <Notification>
-                        <Notification.Icon>
-                            <ErrorStrokeStsIcon20 paletteIndex={1} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                            <Notification.Body.List values={Array.from({ length: 3 }, () => "List item text;")} />
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-            </div>
-        );
+        docs: {
+            source: {
+                code: StatusSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
 export const Business: StoryObj<typeof Notification> = {
-    name: "Business",
+    render: BusinessRender,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: "32px" }}>
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Normal</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Icon>
-                            <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Header>Title text</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                        <Notification.Time time="22:45" />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Button</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Icon>
-                            <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Header>Title text</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Button text
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                        <Notification.Time time="22:45" />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Alert</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Icon>
-                            <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                        </Notification.Icon>
-                        <Notification.Body>
-                            <Notification.Body.Header>Title text</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                This message provides context or highlights important information to note.
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <AlertContext type={EAlertType.INFO}>
-                                    This message provides context or highlights important information to note.
-                                </AlertContext>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                        <Notification.Time time="22:45" />
-                    </Notification>
-                </div>
-            </div>
-        );
+        docs: {
+            source: {
+                code: BusinessSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
 export const BusinessStack: StoryObj<typeof Notification> = {
     name: "Business Stack",
+    render: BusinessStackRender,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: "32px" }}>
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Normal</h3>
-                    <NotificationGrouped>
-                        <Notification withExtraBottomPadding>
-                            <Notification.Icon>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                            </Notification.Icon>
-                            <Notification.Body>
-                                <Notification.Body.Header>Title text</Notification.Body.Header>
-                                <Notification.Body.Content>
-                                    This message provides context or highlights important information to note.
-                                </Notification.Body.Content>
-                            </Notification.Body>
-                            <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                            <Notification.Time time="22:45" />
-                        </Notification>
-                    </NotificationGrouped>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Button</h3>
-                    <NotificationGrouped>
-                        <Notification withExtraBottomPadding>
-                            <Notification.Icon>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                            </Notification.Icon>
-                            <Notification.Body>
-                                <Notification.Body.Header>Title text</Notification.Body.Header>
-                                <Notification.Body.Content>
-                                    This message provides context or highlights important information to note.
-                                </Notification.Body.Content>
-                                <Notification.Body.Footer>
-                                    <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                        Button text
-                                    </Button>
-                                </Notification.Body.Footer>
-                            </Notification.Body>
-                            <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                            <Notification.Time time="22:45" />
-                        </Notification>
-                    </NotificationGrouped>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Alert</h3>
-                    <NotificationGrouped>
-                        <Notification withExtraBottomPadding>
-                            <Notification.Icon>
-                                <DefaulticonStrokePrdIcon20 paletteIndex={5} />
-                            </Notification.Icon>
-                            <Notification.Body>
-                                <Notification.Body.Header>Title text</Notification.Body.Header>
-                                <Notification.Body.Content>
-                                    This message provides context or highlights important information to note.
-                                </Notification.Body.Content>
-                                <Notification.Body.Footer>
-                                    <AlertContext type={EAlertType.INFO}>
-                                        This message provides context or highlights important information to note.
-                                    </AlertContext>
-                                </Notification.Body.Footer>
-                            </Notification.Body>
-                            <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                            <Notification.Time time="22:45" />
-                        </Notification>
-                    </NotificationGrouped>
-                </div>
-            </div>
-        );
+        docs: {
+            source: {
+                code: BusinessStackSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
 export const FeedbackWithoutStars: StoryObj<typeof Notification> = {
     name: "Feedback Without Stars",
+    render: FeedbackWithoutStarsRender,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        const [defaultValue, setDefaultValue] = useState("");
-        const [filledValue, setFilledValue] = useState("Нет возможности экспортировать данные из заказов");
-        const [errorValue, setErrorValue] = useState("");
-
-        const handleDefaultChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-            setDefaultValue(e.target.value);
-        };
-
-        const handleFilledChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-            setFilledValue(e.target.value);
-        };
-
-        const handleErrorChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-            setErrorValue(e.target.value);
-        };
-
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: "32px" }}>
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Default</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Предложите идею</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Чего не хватает вам и что оценят другие пользователи?</div>
-                                <TextField
-                                    className="notificationBodyTextArea"
-                                    status={EFormFieldStatus.DEFAULT}
-                                    inputProps={{
-                                        value: defaultValue,
-                                        onChange: handleDefaultChange,
-                                    }}
-                                    label={"Как мы можем улучшить сервис?"}
-                                />
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Отправить
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Filled</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Предложите идею</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Чего не хватает вам и что оценят другие пользователи?</div>
-                                <TextField
-                                    className="notificationBodyTextArea"
-                                    status={EFormFieldStatus.DEFAULT}
-                                    inputProps={{
-                                        value: filledValue,
-                                        onChange: handleFilledChange,
-                                    }}
-                                    label={undefined}
-                                />
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Отправить
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Error</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Предложите идею</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Чего не хватает вам и что оценят другие пользователи?</div>
-                                <TextField
-                                    className="notificationBodyTextArea"
-                                    status={EFormFieldStatus.ERROR}
-                                    description={
-                                        <Text tag="div" size={ETextSize.B4} type={EFontType.ERROR}>
-                                            Введите текст сообщения.
-                                        </Text>
-                                    }
-                                    inputProps={{
-                                        value: errorValue,
-                                        onChange: handleErrorChange,
-                                    }}
-                                    label={"Как мы можем улучшить сервис?"}
-                                />
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Отправить
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Success</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <div className="finalStatus">
-                                <SuccessgradientStsIcon96 />
-                                <Notification.Body.Header>Спасибо за оценку!</Notification.Body.Header>
-                                <Notification.Body.Content>
-                                    Ваши отзывы помогают нам стать лучше.
-                                </Notification.Body.Content>
-                            </div>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-            </div>
-        );
+        docs: {
+            source: {
+                code: FeedbackWithoutStarsSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
 export const FeedbackWithStars: StoryObj<typeof Notification> = {
     name: "Feedback With Stars",
+    render: FeedbackWithStarsRender,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        const [value, setValue] = useState("");
-        const checkboxes = [
-            "Встречается неточная информация",
-            "Уведомления приходят ночью",
-            "Не хватает данных в сообщении",
-        ];
-
-        const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-            setValue(e.target.value);
-        };
-
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: "32px" }}>
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Default</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Оцените оформление кредита</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Оцените, насколько удобным был процесс оформления недавно полученого кредита.</div>
-                                <div style={{ display: "flex", justifyContent: "center", marginTop: "16px" }}>
-                                    {Array.from({ length: 5 }).map((_, index) => (
-                                        <StarStrokeSrvIcon32
-                                            key={index}
-                                            paletteIndex={0}
-                                            style={{ marginLeft: index !== 0 ? "24px" : "0px" }}
-                                        />
-                                    ))}
-                                </div>
-                            </Notification.Body.Content>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Filled</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Оцените оформление кредита</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Оцените, насколько удобным был процесс оформления недавно полученого кредита.</div>
-                                <div style={{ display: "flex", justifyContent: "center", marginTop: "16px" }}>
-                                    {Array.from({ length: 5 }).map((_, index) => (
-                                        <StarStrokeSrvIcon32
-                                            key={index}
-                                            paletteIndex={0}
-                                            style={{ marginLeft: index !== 0 ? "24px" : "0px" }}
-                                        />
-                                    ))}
-                                </div>
-                                <TextField
-                                    className="notificationBodyTextArea"
-                                    status={EFormFieldStatus.DEFAULT}
-                                    inputProps={{
-                                        value: value,
-                                        onChange: handleChange,
-                                    }}
-                                    label={"Как мы можем улучшить сервис?"}
-                                />
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Отправить
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Error</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <Notification.Body.Header>Оцените оформление кредита</Notification.Body.Header>
-                            <Notification.Body.Content>
-                                <div>Оцените, насколько удобным был процесс оформления недавно полученого кредита.</div>
-                                <div style={{ display: "flex", justifyContent: "center", marginTop: "16px" }}>
-                                    {Array.from({ length: 5 }).map((_, index) => (
-                                        <StarStrokeSrvIcon32
-                                            key={index}
-                                            paletteIndex={0}
-                                            style={{ marginLeft: index !== 0 ? "24px" : "0px" }}
-                                        />
-                                    ))}
-                                </div>
-                                <CheckboxYGroup
-                                    aria-labelledby="checkbox-y-group-label"
-                                    className="notificationBodyCheckbox"
-                                >
-                                    {checkboxes.map((value, index) => (
-                                        <Checkbox key={index} name="checkbox-y-group">
-                                            {value}
-                                        </Checkbox>
-                                    ))}
-                                </CheckboxYGroup>
-                            </Notification.Body.Content>
-                            <Notification.Body.Footer>
-                                <Button theme={EButtonTheme.SECONDARY} size={EComponentSize.SM}>
-                                    Отправить
-                                </Button>
-                            </Notification.Body.Footer>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-
-                <div>
-                    <h3 style={{ marginBottom: "16px", fontSize: "18px", fontWeight: "600" }}>Success</h3>
-                    <Notification withExtraBottomPadding>
-                        <Notification.Body>
-                            <div className="finalStatus">
-                                <SuccessgradientStsIcon96 />
-                                <Notification.Body.Header>Спасибо за оценку!</Notification.Body.Header>
-                                <Notification.Body.Content>
-                                    Ваши отзывы помогают нам стать лучше.
-                                </Notification.Body.Content>
-                            </div>
-                        </Notification.Body>
-                        <Notification.Close onClick={() => alert("Обработчик закрытия нотификации")} />
-                    </Notification>
-                </div>
-            </div>
-        );
+        docs: {
+            source: {
+                code: FeedbackWithStarsSource,
+                language: "tsx",
+            },
+        },
     },
 };
