+++ stories/release-notes/v1/1.33.0.mdx
@@ -50,6 +50,9 @@ import { Meta, Title, Heading } from "@storybook/addon-docs/blocks";
 **Notification**
 
 - Исправлен шрифт заголовка.
+- Компоненты `Notification`, `NotificationGrouped` и `Notification.Body` переведены на использование `forwardRef` (ref ведёт на корневой `div`). **Breaking**: изменился ref-контракт.
+- Удалены типы `INotificationSFC` и `INotificationBodySFC`. **Breaking**: тип `INotificationSFC` больше не экспортируется из пакета.
+- Тип `Notification.Time.time` расширен с `React.ReactText` до `React.ReactNode`.
 
 **Stepper**
